+++ ModelCatalogueCorePluginTestApp/grails-app/services/org/modelcatalogue/core/catalogueelement/UserCatalogueElementService.groovy
@@ -3,7 +3,7 @@ package org.modelcatalogue.core.catalogueelement
 import grails.util.GrailsNameUtils
 import groovy.transform.CompileStatic
 import org.modelcatalogue.core.CatalogueElement
-import org.modelcatalogue.core.security.UserGormService
+import org.modelcatalogue.core.persistence.UserGormService
 import org.modelcatalogue.core.security.User
 
 @CompileStatic
