+++ ModelCatalogueCorePluginTestApp/grails-app/services/org/modelcatalogue/core/security/ss2x/SpringSecurity2SecurityService.groovy
@@ -15,9 +15,9 @@ import org.modelcatalogue.core.security.DataModelAclService
 import org.modelcatalogue.core.security.MetadataRolesUtils
 import org.modelcatalogue.core.security.Role
 import org.modelcatalogue.core.security.User
-import org.modelcatalogue.core.security.UserGormService
 import org.modelcatalogue.core.security.UserRole
 import org.springframework.security.acls.domain.BasePermission
+import org.modelcatalogue.core.persistence.UserGormService
 import org.springframework.security.core.Authentication
 import org.springframework.security.web.authentication.logout.LogoutHandler
 import javax.servlet.http.HttpServletRequest
